+++ NEWS.md
@@ -7,3 +7,8 @@
 * First public release
 * Tools streamlined (non-git-related tools culled)
 * gitcreds package adopted as optional way of handling credentials
+
+# git4r 0.1.2
+
+* Bug fixes where you must have valid git username before a pull
+* Non-bare option removed for a shared-drive remote (git2r::push limitations)